refactor(registry): lettore consolidato in toolkit/registry/ (home unica: builder+schemi+reader)
