+++ toolkit/registry/__init__.py
@@ -1,9 +1,11 @@
-"""Registry builder condiviso DataCivicLab.
+"""Registry builder + lettore condiviso DataCivicLab.
 
-Genera gli artifact registry del Lab — clean_catalog, pipeline_signals,
-mart_catalog — riusando il runtime del toolkit (config model, path resolver,
-run_state, parquet_schema) con layout e path contract parametrizzati, così ogni
-repo (dataset-incubator, eurostat, dcl-bologna, ...) riusa la stessa logica.
+Capacità unica del Lab per gli artifact registry:
+- ``builders``: genera clean_catalog, mart_catalog, pipeline_signals, codelists
+  (riusando il runtime del toolkit: config model, path resolver, run_state,
+  parquet_schema) con layout e path contract parametrizzati.
+- ``reader``: consulta gli artifact committati nei repo del workspace
+  (CLI ``toolkit registry`` e MCP ``toolkit_registry_*``).
 
 Chiave canonica: ``dataset.name`` (underscore) — le directory del repo sono
 solo contenitori per trovare il dataset.yml, mai identità del dataset.
@@ -11,6 +13,7 @@ solo contenitori per trovare il dataset.yml, mai identità del dataset.
 
 from toolkit.registry.layout import DatasetManifest, RepoLayout, iter_manifests, load_manifest
 from toolkit.registry.paths import PathContract
+from toolkit.registry.reader import list_registries, show_registry
 from toolkit.registry.builders import (
     build_clean_catalog,
     build_codelists,
@@ -29,5 +32,7 @@ __all__ = [
     "build_registry",
     "build_signals",
     "iter_manifests",
+    "list_registries",
     "load_manifest",
+    "show_registry",
 ]
